--- a/timelock-server/src/main/java/com/palantir/atlasdb/timelock/paxos/servlets/DelegatingServlet.java
+++ b/timelock-server/src/main/java/com/palantir/atlasdb/timelock/paxos/servlets/DelegatingServlet.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Copyright 2017 Palantir Technologies, Inc. All rights reserved.
  *
  * Licensed under the BSD-3 License (the "License");