--- a/timelock-server/src/main/java/com/palantir/atlasdb/timelock/paxos/servlets/ServletFactory.java
+++ b/timelock-server/src/main/java/com/palantir/atlasdb/timelock/paxos/servlets/ServletFactory.java
@@ -1,0 +1,50 @@
+/*
+ * Copyright 2017 Palantir Technologies, Inc. All rights reserved.
+ *
+ * Licensed under the BSD-3 License (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://opensource.org/licenses/BSD-3-Clause
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.palantir.atlasdb.timelock.paxos.servlets;
+
+import javax.ws.rs.HttpMethod;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.palantir.leader.PingableLeader;
+import com.palantir.paxos.PaxosAcceptor;
+import com.palantir.timestamp.TimestampService;
+
+public final class ServletFactory {
+    private ServletFactory() {
+        // factory, so no.
+    }
+
+    public static DelegatingServlet<Boolean> leaderPingServlet(PingableLeader pingable) {
+        return new DelegatingServlet<>(pingable::ping, HttpMethod.GET, new TextPlainResponseStrategy<>());
+    }
+
+    public static DelegatingServlet<Long> getLatestSequencePreparedOrAcceptedServlet(
+            ObjectMapper mapper, PaxosAcceptor acceptor) {
+        return new DelegatingServlet<>(
+                acceptor::getLatestSequencePreparedOrAccepted,
+                HttpMethod.POST,
+                new ApplicationJsonResponseStrategy<>(mapper));
+    }
+
+    public static DelegatingServlet<Long> freshTimestampServlet(
+            ObjectMapper mapper, TimestampService timestampService) {
+        return new DelegatingServlet<>(
+                timestampService::getFreshTimestamp,
+                HttpMethod.POST,
+                new ApplicationJsonResponseStrategy<>(mapper));
+    }
+}