--- a/timelock-server/src/main/java/com/palantir/atlasdb/timelock/paxos/servlets/ServletRegistration.java
+++ b/timelock-server/src/main/java/com/palantir/atlasdb/timelock/paxos/servlets/ServletRegistration.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2017 Palantir Technologies, Inc. All rights reserved.
+ *
+ * Licensed under the BSD-3 License (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://opensource.org/licenses/BSD-3-Clause
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.palantir.atlasdb.timelock.paxos.servlets;
+
+import javax.servlet.http.HttpServlet;
+
+import org.eclipse.jetty.servlet.ServletHolder;
+
+import io.dropwizard.setup.Environment;
+
+public final class ServletRegistration {
+    private ServletRegistration() {
+        // no
+    }
+
+    public static void registerServlet(Environment environment, HttpServlet servlet, String path) {
+        environment.getApplicationContext().addServlet(new ServletHolder(servlet), path);
+    }
+}