Throwing this out for initial ideas/review. This is still a WIP though the "core" methodology is likely to be pretty similar to what's here.

**TODO**

- [ ] Some refactoring/reorganisation to ensure this stuff is cleaner
- [ ] Unit tests
- [ ] Benchmark on a remote three node cluster
- [ ] Release notes - perf jumped by foo, from step 2

**Goals (and why)**:
See https://github.com/palantir/atlasdb/issues/1836#issuecomment-297207656 - specifically, this part.

```
-- SSL / HTTP2 -- 
websocket: 110 us

feign + jersey: 310 us
(...)
feign + jetty: 245 us
```

JAX-RS/Jersey dispatch does introduce some overheads. Apparently HTTP/2 overhead over websockets itself isn't too bad; the overhead is largely introduced through Feign and Jersey.

This change switches us to use raw Jetty servlets for the three most common calls on TimeLock:

* leader/ping
* timestamp/fresh-timestamp
* acceptor/getLatestSequencePreparedOrAccepted (for both the leadership and timestamp bound acceptors)

Locally, testing timestamp/fresh-timestamp yielded a 25% improvement (280 --> 210 us), which is not too different from @nziebart's results above. Firing 8 threads which hammered timestamp/fresh-timestamp yielded a much larger improvement (~45%), as well.

Stage 2 could be for us to devise a high performance client to possibly replace the Feign/OkHttp configuration we have in `AtlasDbHttpClients`. That'll probably be a riskier and more invasive change than this one, though it could yield another (probably bigger) perf win.

**Implementation Description (bullets)**:

* Introduce a fair bit of servlet machinery in the timelock-server project, which is a framework for delegating Java functions to servlets.
* Register HTTP servlets for the functions defined above, which take precedence over Jersey resolution.

**Concerns (what feedback would you like?)**:

* Are we handling errors correctly even when this switch is done?
* Is the complexity worth it?

**Where should we start reviewing?**: `DelegatingServlet` is probably the best start point.

**Priority (whenever / two weeks / yesterday)**: for now, whenever; though if/when timelock perf becomes a concern this may subsequently increase

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/palantir/atlasdb/1904)
<!-- Reviewable:end -->