@schlosna 2017-05-15
A couple thoughts:

If we go pure servlet route, might be worth testing timelock running on [undertow](http://undertow.io/) to reduce raw latency & improve throughput.

Would prefer not to have to go this route, but we have an implementation of time + lock on Netty raw TCP + TLS + protobuf.