@jeremyk-91 2017-05-16
@schlosna Agreed. From the usage patterns we've observed these 3 cover >90% of the requests, so this would be a good starting point, though of course there's scope to push further as you've outlined.

I think the general direction we're going in for scaling out to _lots_ of services is actually to run multiple timelocks, so we hopefully won't have to go the full netty route.