@nziebart 2017-06-04
I think this is definitely worth exploring, as my own rough IL benchmarks from a while back showed significant gains from this approach (see below - note these were returning a constant timestamp value to isolate the http / paxos overhead).

There may also be some gains to be had from eliminating gzip and looking at using GET instead of POST (okhttp does an extra check on the connection with a 1ms timeout on POST requests - see https://github.com/square/okhttp/blob/e14817d993c39c8a80747179bf8295b15d36372a/okhttp/src/main/java/okhttp3/internal/connection/RealConnection.java#L497).

We should probably prioritize getting some reliable benchmarking set up for distributed a Timelock cluster to better evaluate these things.

```
jersey (current impl):
{
  "average": 1.0512964138,
  "p99": 1.80686,
  "totalTime": 5258.364541,
  "requestsPerClient": 5000,
  "p50": 0.995959,
  "numClients": 1,
  "throughput": 950.8659890379404,
  "p95": 1.152333
}
{
  "average": 1.807613482,
  "p99": 3.914599,
  "totalTime": 9052.980664,
  "requestsPerClient": 5000,
  "p50": 1.70762,
  "numClients": 10,
  "throughput": 5523.042835916965,
  "p95": 2.299369
}

direct jetty servlet:
{
  "average": 0.7818532290000001,
  "p99": 0.943098,
  "totalTime": 3911.089999,
  "requestsPerClient": 5000,
  "p50": 0.732051,
  "numClients": 1,
  "throughput": 1278.4159917768234,
  "p95": 0.840117
}
{
  "average": 1.35009940252,
  "p99": 3.058213,
  "totalTime": 6763.610948,
  "requestsPerClient": 5000,
  "p50": 1.257434,
  "numClients": 10,
  "throughput": 7392.500896992752,
  "p95": 1.596452
}

websockets:
{
  "average": 0.6206893796,
  "p99": 0.727864,
  "totalTime": 3106.369697,
  "requestsPerClient": 5000,
  "p50": 0.601137,
  "numClients": 1,
  "throughput": 1609.5959231217032,
  "p95": 0.673148
}
{
  "average": 0.9883441079199999,
  "p99": 1.268547,
  "totalTime": 4964.267132,
  "requestsPerClient": 5000,
  "p50": 0.909335,
  "numClients": 10,
  "throughput": 10071.980147421285,
  "p95": 1.065284
}
```