@hsaraogi 2017-06-12
This a proof-of-concept for future reference for timelock perf improvements.